backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
PyJWT==2.8.0
google-auth-oauthlib==1.0.0
google-auth==2.23.0
google-api-python-client==2.104.0